Updated both versions of input_creation
